Fix incorrect scoring project test
Given the rules as described, the original first assertion in
`test-score-of-mixed-is-sum` should have returned 50 (one five, no
sequences of threes, no ones).

Assuming that the intent of the test case is to demonstrate that
numbers other than one and five are worth zero points but the values
of ones and fives are summed, this patch changes the final three to a
one, resulting in a score of 150.