Update evaluation.lsp
The material referenced is no longer available directly, instead there is a link to a google drive archive. Alternatively, the book is available on amazon. I have linked to the original website with the link to amazon, in case the author decides to update the link or use a different publisher.